refactor(ChildProcesses): :construction: work in progress
